requirements.txt
addict==2.4.0
appdirs==1.4.4
asttokens==2.0.5
attr==0.3.2
attrs==23.2.0
Automat==22.10.0
backcall==0.2.0
Brotli==1.0.9
buildtools==1.0.6
cachetools==5.3.3
causal-conv1d==1.1.3.post1
certifi==2024.2.2
charset-normalizer==3.3.2
cityscapesScripts==2.2.2
cloudpickle==3.0.0
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
decorator==5.1.1
docopt==0.6.2
einops==0.7.0
executing==0.8.3
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.10.0
furl==2.1.3
future==0.18.3
fvcore==0.1.5.post20221221
greenlet==3.0.3
huggingface-hub==0.17.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
incremental==22.10.0
iopath==0.1.10
ipython==8.12.2
jedi==0.18.1
Jinja2==3.1.3
kiwisolver==1.4.5
mamba-ssm==1.1.3.post1
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.6
MedPy==0.4.0
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mmcv-full==1.7.2
ninja==1.11.1.1
numpy==1.23.5
nvidia-ml-py==12.535.133
nvitop==1.3.2
opencv-python==4.5.1.48
orderedmultidict==1.0.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
platformdirs==4.1.0
portalocker==2.8.2
prettytable==3.10.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0.1
redo==2.0.4
regex==2023.12.25
requests==2.31.0
safetensors==0.4.0
scipy==1.10.1
setuptools==68.2.2
SimpleITK==2.3.1
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.27
stack-data==0.2.0
submitit==1.5.1
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
timm==0.9.12
tokenizers==0.15.2
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchsummary==1.5.1
torchvision==0.13.1
tqdm==4.65.0
traitlets==5.7.1
transformers==4.37.2
triton==2.2.0
Twisted==23.10.0
typing==3.7.4.3
typing_extensions==4.7.1
urllib3==2.2.0
wcwidth==0.2.5
wheel==0.41.2
yacs==0.1.8
yapf==0.40.1
zipp==3.17.0
zope.interface==6.2
